Added coverage test: truffle run coverage
tests seem to not pass right now (Marketplace deployment runs out of gas because of coverage instrumentation)
